Add Logout and Status commands
tfx status will print out the status of the currently logged in connection, if there is one.

tfx logout will logout of the currently logged in connection if there is one. Logging out does not remove the cached credentials for that collection. This allows the user to continue using that collection by opting in on the command line with --collectionurl and using the cached credentials. It just causes server commands to not use a specific collection by default.

These two commands can be used in conjunction to help manage working with multiple accounts/collections.