Uses the manifest's contributions to find task.json
Changes the way task.json files are looked up. This fixes a number of scenarios where a task.json would not be detected.

This also improves performance since it significantly reduces the number of folders to scan from all child folders to just the contribution folder and its direct children.

This enabled checking for supported runner versions across all versions of a task, not logging a warning if an older task version, included for on-premise backwards compatibility, doesn't support a non-deprecated runner when a later version does.