Upgrade tmp package and refactor json validation functions (#508)
* Upgrading tmp to address security vulnerability

* Refactor json validation functions to remove 'this' context and improve clarity; enhance warning message for deprecated runners with version info.

* Bump version to 0.21.3 in package.json

* Improve warning message for deprecated runners to include version information

Print ? when a version is missing from the manifest.

* Implemented proper backwards compatibility checking for runners.

Only log a warning is none of the packaged task versions support a non-deprecated runner.

* Uses the manifest's contributions to find task.json

Changes the way task.json files are looked up. This fixes a number of scenarios where a task.json would not be detected.

This also improves performance since it significantly reduces the number of folders to scan from all child folders to just the contribution folder and its direct children.

This enabled checking for supported runner versions across all versions of a task, not logging a warning if an older task version, included for on-premise backwards compatibility, doesn't support a non-deprecated runner when a later version does.

* Removes redundant comment

* Removing unused methods left over from refactoring.